Revert "Added SpringBoot and JPA dependencies. Added Resource folder. Added Controller and Service packages."
