+++ README.md
@@ -2,12 +2,12 @@
 
 ---
 
-### About
+###About
 
 ---
 The purpose of the system is to allow a user to book parking space in advance on a system for a specified amount of time.  
 This system is used by the CPUT registered students, lecturers and guests. 
-## Current UML of the System
+##Current UML of the System
 
 ---
 ![](uml/ParkingLotSystemUMLdraft5.PNG)
