--- src/main/java/za/ac/cput/carparkingservicesystem/MyApplication.java
@@ -1,12 +1,0 @@
-package za.ac.cput.carparkingservicesystem;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MyApplication {
-
-    public static void main(String[] args){
-        SpringApplication.run(MyApplication.class, args);
-    }
-}
