--- src/main/java/za/ac/cput/carparkingservicesystem/controller/ControllerClass.java
@@ -1,7 +1,0 @@
-/**
- * To be removed by team lead
- */
-package za.ac.cput.carparkingservicesystem.controller;
-
-public class ControllerClass {
-}
