--- src/main/java/za/ac/cput/carparkingservicesystem/factory/FactoryClass.java
@@ -1,7 +1,0 @@
-/**
- * To be removed by team lead.
- */
-package za.ac.cput.carparkingservicesystem.factory;
-
-public class FactoryClass {
-}
