--- src/main/java/za/ac/cput/carparkingservicesystem/factory/ParkingSpaceFactory.java
@@ -1,22 +1,0 @@
-/**
- * ParkingSpaceFactory.java
- * Factory for ParkingSpace
- * Author: Sonwabile Gxoyiya (219267189)
- * Date: 5 April 2022
- */
-package za.ac.cput.carparkingservicesystem.factory;
-
-import za.ac.cput.carparkingservicesystem.entity.ParkingSpace;
-import za.ac.cput.carparkingservicesystem.helper.GenerateID;
-
-public class ParkingSpaceFactory {
-    public static ParkingSpace build( String parkingSpaceNumber, String parkingLotID, boolean isEmpty){
-        String parkingSpaceID = GenerateID.generateUUID();
-        return new ParkingSpace.Builder()
-                .setParkingSpaceID(parkingSpaceID)
-                .setParkingSpaceNumber(parkingSpaceNumber)
-                .setParkingLotID(parkingLotID)
-                .setIsEmpty(isEmpty)
-                .build();
-    }
-}
