--- src/main/java/za/ac/cput/carparkingservicesystem/repository/ICampusRepository.java
@@ -1,16 +1,0 @@
-/**
- * ICampusRepository.java
- * Interface for other classes to extend
- * Author: Sonwabile Gxoyiya (219267189)
- * Date: 9 April 2022
- */
-package za.ac.cput.carparkingservicesystem.repository;
-
-import java.util.List;
-
-import za.ac.cput.carparkingservicesystem.entity.Campus;
-
-public interface ICampusRepository extends IRepository<Campus, String>{
-
-    public List<Campus> getAll();
-}
