--- src/main/java/za/ac/cput/carparkingservicesystem/repository/RepositoryClass.java
@@ -1,7 +1,0 @@
-/**
- * To be removed by team lead.
- */
-package za.ac.cput.carparkingservicesystem.repository;
-
-public class RepositoryClass {
-}
