--- src/main/java/za/ac/cput/carparkingservicesystem/service/ServiceClass.java
@@ -1,7 +1,0 @@
-/**
- * To be removed by team lead
- */
-package za.ac.cput.carparkingservicesystem.service;
-
-public class ServiceClass {
-}
