+++ src/main/java/za/ac/cput/entity/EntityClass.java
@@ -1,0 +1,7 @@
+/**
+ * To be removed by team lead.
+ */
+package za.ac.cput.entity;
+
+public class EntityClass {
+}
