+++ src/main/java/za/ac/cput/entity/ParkingSpace.java
@@ -1,0 +1,87 @@
+/**
+ * ParkingSpace.java
+ * Entity for ParkingSpace
+ * Author: Sonwabile Gxoyiya (219267189)
+ * Date: 5 April 2022
+ */
+package za.ac.cput.entity;
+
+public class ParkingSpace {
+    private String parkingSpaceID;
+    private String parkingSpaceNumber;
+    private String parkingLotID;
+    private boolean isEmpty;
+
+    public String getParkingSpaceID(){
+        return parkingSpaceID;
+    }
+
+    public String getParkingSpaceNumber(){
+        return parkingSpaceNumber;
+    }
+
+    public String getParkingLotID(){
+        return parkingLotID;
+    }
+
+    public boolean getIsEmpty(){return isEmpty;}
+
+    private ParkingSpace(Builder builder){
+        this.parkingSpaceID = builder.parkingSpaceID;
+        this.parkingSpaceNumber = builder.parkingSpaceNumber;
+        this.parkingLotID = builder.parkingLotID;
+        this.isEmpty = builder.isEmpty;
+    }
+
+    @Override
+    public String toString() {
+        return "ParkingSpace{" +
+                "parkingSpaceID='" + parkingSpaceID + '\'' +
+                ", parkingSpaceNumber='" + parkingSpaceNumber + '\'' +
+                ", parkingLotID='" + parkingLotID + '\'' +
+                ", isEmpty=" + isEmpty +
+                '}';
+    }
+
+    //builder Section
+
+    public static class Builder{
+        private String parkingSpaceID;
+        private String parkingSpaceNumber;
+        private String parkingLotID;
+        private boolean isEmpty;
+
+        public Builder setParkingSpaceID(String parkingSpaceID){
+            this.parkingSpaceID = parkingSpaceID;
+            return this;
+        }
+
+        public Builder setParkingSpaceNumber(String parkingSpaceNumber){
+            this.parkingSpaceNumber = parkingSpaceNumber;
+            return this;
+        }
+
+        public Builder setParkingLotID(String parkingLotID){
+            this.parkingLotID = parkingLotID;
+            return this;
+        }
+
+        public Builder setIsEmpty(boolean isEmpty){
+            this.isEmpty = isEmpty;
+            return this;
+        }
+
+        public Builder copy(ParkingSpace parkingSpace){
+            this.parkingSpaceID = parkingSpace.parkingSpaceID;
+            this.parkingSpaceNumber = parkingSpace.parkingSpaceNumber;
+            this.parkingLotID = parkingSpace.parkingLotID;
+            this.isEmpty = parkingSpace.isEmpty;
+            return this;
+
+        }
+
+        public ParkingSpace build(){
+            return new ParkingSpace(this);
+        }
+    }
+}
