+++ src/main/java/za/ac/cput/factory/FactoryClass.java
@@ -1,0 +1,7 @@
+/**
+ * To be removed by team lead.
+ */
+package za.ac.cput.factory;
+
+public class FactoryClass {
+}
