+++ src/main/java/za/ac/cput/repository/ICampusRepository.java
@@ -1,0 +1,15 @@
+/**
+ * ICampusRepository.java
+ * Interface for other classes to extend
+ * Author: Sonwabile Gxoyiya (219267189)
+ * Date: 9 April 2022
+ */
+package za.ac.cput.repository;
+
+import java.util.List;
+import za.ac.cput.entity.Campus;
+
+public interface ICampusRepository extends IRepository<Campus, String>{
+
+    public List<Campus> getAll();
+}
