+++ src/main/java/za/ac/cput/repository/IRepository.java
@@ -1,0 +1,15 @@
+/**
+ * IRepository.java
+ * Interface for other classes to extend
+ * Author: Sonwabile Gxoyiya (219267189)
+ * Date: 9 April 2022
+ */
+package za.ac.cput.repository;
+
+public interface IRepository<T, ID> {
+
+    T create(T t);
+    T read(ID id);
+    T update(T t);
+    boolean delete(ID id);
+}
