+++ src/main/java/za/ac/cput/repository/RepositoryClass.java
@@ -1,0 +1,7 @@
+/**
+ * To be removed by team lead.
+ */
+package za.ac.cput.repository;
+
+public class RepositoryClass {
+}
