--- src/main/resources/application.yaml
@@ -1,11 +1,0 @@
-spring:
-  jpa:
-    hibernate:
-      ddl-auto: create
-      show-sql: true
-
-
-  datasource:
-    url: jdbc:mysql://localhost:3306/carparkingservice-db?createDatabaseIfNotExist=true
-    username: root
-    password: password
