--- src/test/java/za/ac/cput/carparkingservicesystem/controller/ControllerClassTest.java
@@ -1,10 +1,0 @@
-/**
- * To be removed by team lead
- */
-package za.ac.cput.carparkingservicesystem.controller;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class ControllerClassTest {
-
-}
