--- src/test/java/za/ac/cput/carparkingservicesystem/factory/FactoryClassTest.java
@@ -1,12 +1,0 @@
-/**
- * To be removed by team lead.
- */
-package za.ac.cput.carparkingservicesystem.factory;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class FactoryClassTest {
-
-}
