+++ src/test/java/za/ac/cput/factory/FactoryClassTest.java
@@ -1,0 +1,12 @@
+/**
+ * To be removed by team lead.
+ */
+package za.ac.cput.factory;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class FactoryClassTest {
+
+}
