+++ src/test/java/za/ac/cput/repository/RepositoryClassTest.java
@@ -1,0 +1,10 @@
+/**
+ * To be removed by team lead.
+ */
+package za.ac.cput.repository;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class RepositoryClassTest {
+
+}
